requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
hypothesis==6.100.1
pytest==8.2.0
httpx==0.27.0
google-generativeai==0.5.4
pytest-asyncio==0.23.6
anyio==4.3.0
